requirements.txt
Django==1.6.2
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Pillow==2.3.1
Pygments==1.6
South==0.8.4
Werkzeug==0.9.4
argparse==1.2.1
boto==2.27.0
bpython==0.12
dj-database-url==0.3.0
dj-static==0.0.5
django-appconf==0.6
django-bootstrap3==3.2.1
django-extensions==1.3.3
django-imagekit==3.2
django-photologue==2.7
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==18.0
itsdangerous==0.23
newrelic==2.16.0.12
pilkit==1.1.6
psycopg2==2.5.2
pystache==0.5.3
six==1.6.1
static==1.0.2
wsgiref==0.1.2
